Allow to control boot
The boot process does load default data, load module, and start
receiving broadcasted events, which is a problem for dev.exs, test, and
seedind data that only needs to insert data.

So it introduces a new key :skip_boot? in config to allow disabling boot
in some specific scenarios.

Also moved the actual site config to Beacon.Config instead
SiteSupervisor so we can update the values.